+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.0+3
+
+Updated the base WhatsApp Sticker Libraries
+
+Fixed an issue where dynamic sticker packs where not being added to WhatsApp
+
+Added a new method to notify changes to sticker packs
+
+Rebuilt the Example App.  It now has two example sticker packs and displays more information
+
+Added a new example app for dynamic content showcasing how to use dynamic sticker packs
+
 ## 1.0.0+2
 
 Upgraded the WhatsApp Stickers Extentions for Android
@@ -7,15 +19,19 @@ Upgraded the WhatsApp Stickers Extentions for Android
 Bumped this straight to version 1.0.0
 
 Created a listener for the`addStickerPack` method
+
 Added a new build config field for the sticker_packs.json file
+
 Added propper documentation to the plugin dart file
 
 Updated the plugin import statements in the example app
+
 Added a title to example app
 
 Updated the readme
+
 Updated .gitignore
 
 ## 0.0.1
 
-* Initial Release
+Initial Release